Произведён рефакторинг, учитывающий комментарии ментора, стиль кода Google.
