+++ CMakeLists.txt
@@ -7,7 +7,7 @@ add_compile_options(-pedantic -Wall -Wextra -Wshadow -Werror)
 
 find_package(Curses REQUIRED)
 
-add_executable(CppDiagHelper src/main.cpp src/TextUI.cpp)
+add_executable(CppDiagHelper src/main.cc src/ui.cc)
 target_compile_options(CppDiagHelper PRIVATE "$<$<CONFIG:Debug>:-fsanitize=address,undefined,leak>")
 set(
 	CMAKE_EXE_LINKER_FLAGS_DEBUG
