+++ src/main.cc
@@ -1,0 +1,9 @@
+#include <cstdio>
+#include "ui.h"
+
+using diags::ui::Console;
+
+int main() {
+	Console console;
+	return 0;
+}
