--- src/main.cpp
@@ -1,8 +1,0 @@
-#include <cstdio>
-#include "TextUI.hpp"
-
-int main() {
-	TextUI::initConsole();
-	TextUI::deinitConsole();
-	return 0;
-}
